Includes changes from later branch.
