+++ packages/evo-objects/src/evo/objects/typed/regular_grid.py
@@ -11,32 +11,27 @@
 
 from __future__ import annotations
 
-import sys
 from dataclasses import dataclass
 
 import numpy as np
 import pandas as pd
 from pydantic import TypeAdapter
 
-from evo.common import IContext, IFeedback
+from evo.common import IFeedback
 from evo.common.utils import NoFeedback
 from evo.objects import SchemaVersion
 
 from ._adapters import AttributesAdapter
-from .base import BaseSpatialObject, BaseSpatialObjectData, DatasetProperty, SchemaProperty
+from ._property import SchemaProperty
+from .base import BaseSpatialObject, BaseSpatialObjectData, ConstructableObject, DatasetProperty
 from .dataset import Dataset
+from .exceptions import ObjectValidationError
 from .types import BoundingBox, Point3, Rotation, Size3d, Size3i
-
-if sys.version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
 
 __all__ = [
     "Cells",
     "Regular3DGrid",
     "Regular3DGridData",
-    "SizeChangeError",
     "Vertices",
 ]
 
@@ -79,25 +74,18 @@ class Regular3DGridData(BaseSpatialObjectData):
     rotation: Rotation | None = None
 
     def __post_init__(self):
-        if self.cell_data is not None and self.cell_data.shape[0] != self.size.nx * self.size.ny * self.size.nz:
-            raise ValueError(
+        if self.cell_data is not None and self.cell_data.shape[0] != self.size.total_size:
+            raise ObjectValidationError(
                 f"The number of rows in the cell_data dataframe ({self.cell_data.shape[0]}) does not match the number of cells in the grid ({self.size.nx * self.size.ny * self.size.nz})."
             )
-        if self.vertex_data is not None and self.vertex_data.shape[0] != (self.size.nx + 1) * (self.size.ny + 1) * (
-            self.size.nz + 1
-        ):
-            raise ValueError(
+        vertices_expected_length = (self.size.nx + 1) * (self.size.ny + 1) * (self.size.nz + 1)
+        if self.vertex_data is not None and self.vertex_data.shape[0] != vertices_expected_length:
+            raise ObjectValidationError(
                 f"The number of rows in the vertex_data dataframe ({self.vertex_data.shape[0]}) does not match the number of vertices in the grid ({self.size.nx * self.size.ny * self.size.nz})."
             )
 
     def compute_bounding_box(self) -> BoundingBox:
         return _calculate_bounding_box(self.origin, self.size, self.cell_size, self.rotation)
-
-
-class SizeChangeError(Exception):
-    """Exception raised when the size of the grid cannot be changed due to existing attributes in the datasets."""
-
-    pass
 
 
 class Cells(Dataset):
@@ -106,14 +94,21 @@ class Cells(Dataset):
     The order of the cells is assumed to be in z-fastest order.
     """
 
-    size: Size3i
+    size: Size3i = SchemaProperty(
+        "size",
+        TypeAdapter(Size3i),
+    )
 
     async def set_dataframe(self, df: pd.DataFrame, fb: IFeedback = NoFeedback) -> None:
-        if df.shape[0] != self.size.nx * self.size.ny * self.size.nz:
-            raise SizeChangeError(
-                f"The number of rows in the dataframe ({df.shape[0]}) does not match the number of cells in the grid ({self.size.nx * self.size.ny * self.size.nz})."
+        expected_length = self.size.total_size
+        if df.shape[0] != expected_length:
+            raise ObjectValidationError(
+                f"The number of rows in the dataframe ({df.shape[0]}) does not match the number of cells in the grid ({expected_length})."
             )
-        await super().set_dataframe(df, fb)
+        await super().set_dataframe(df, fb=fb)
+
+    def validate(self) -> None:
+        self._check_length(self.size.total_size)
 
 
 class Vertices(Dataset):
@@ -122,17 +117,29 @@ class Vertices(Dataset):
     The order of the cells is assumed to be in z-fastest order.
     """
 
-    size: Size3i
+    _grid_size: Size3i = SchemaProperty(
+        "size",
+        TypeAdapter(Size3i),
+    )
 
-    def set_dataframe(self, df: pd.DataFrame, fb: IFeedback = NoFeedback) -> None:
-        if df.shape[0] != self.size.nx * self.size.ny * self.size.nz:
-            raise SizeChangeError(
-                f"The number of rows in the dataframe ({df.shape[0]}) does not match the number of vertices in the grid ({self.size.nx * self.size.ny * self.size.nz})."
+    @property
+    def size(self) -> Size3i:
+        grid_size = self._grid_size
+        return Size3i(nx=grid_size.nx + 1, ny=grid_size.ny + 1, nz=grid_size.nz + 1)
+
+    async def set_dataframe(self, df: pd.DataFrame, fb: IFeedback = NoFeedback) -> None:
+        expected_length = self.size.total_size
+        if df.shape[0] != expected_length:
+            raise ObjectValidationError(
+                f"The number of rows in the dataframe ({df.shape[0]}) does not match the number of vertices in the grid ({expected_length})."
             )
-        super().set_dataframe(df, fb)
+        await super().set_dataframe(df, fb=fb)
+
+    def validate(self) -> None:
+        self._check_length(self.size.total_size)
 
 
-class Regular3DGrid(BaseSpatialObject):
+class Regular3DGrid(BaseSpatialObject, ConstructableObject[Regular3DGridData]):
     """A GeoscienceObject representing a regular 3D grid.
 
     The object contains a dataset for both the cells and the vertices of the grid.
@@ -140,6 +147,8 @@ class Regular3DGrid(BaseSpatialObject):
     Each of these datasets only contain attribute columns. The actual geometry of the grid is defined by
     the properties: origin, size, cell_size, and rotation.
     """
+
+    _data_class = Regular3DGridData
 
     sub_classification = "regular-3d-grid"
     creation_schema_version = SchemaVersion(major=1, minor=3, patch=0)
@@ -150,7 +159,7 @@ class Regular3DGrid(BaseSpatialObject):
         attributes_adapters=[
             AttributesAdapter(min_major_version=1, max_major_version=1, attribute_list_path="cell_attributes")
         ],
-        data_attribute="cell_data",
+        extract_data=lambda data: data.cell_data,
     )
     vertices: Vertices = DatasetProperty(
         Vertices,
@@ -158,7 +167,7 @@ class Regular3DGrid(BaseSpatialObject):
         attributes_adapters=[
             AttributesAdapter(min_major_version=1, max_major_version=1, attribute_list_path="vertex_attributes")
         ],
-        data_attribute="vertex_data",
+        extract_data=lambda data: data.vertex_data,
     )
     origin: Point3 = SchemaProperty(
         "origin",
@@ -169,21 +178,6 @@ class Regular3DGrid(BaseSpatialObject):
         TypeAdapter(Size3i),
     )
 
-    @size.pre_set
-    def _check_size_on_datasets(self, value: Size3i) -> None:
-        if value != self.cells.size and self.cells.has_attributes():
-            raise SizeChangeError("The size property can only be changed if the cells dataset has no attributes.")
-        if (
-            value != Size3i(nx=self.vertices.size.nx - 1, ny=self.vertices.size.ny - 1, nz=self.vertices.size.nz - 1)
-            and self.vertices.has_attributes()
-        ):
-            raise SizeChangeError("The size property can only be changed if the vertices dataset has no attributes.")
-
-    @size.post_set
-    def _set_size_on_datasets(self, value: Size3i) -> None:
-        self.cells.size = value
-        self.vertices.size = Size3i(nx=value.nx + 1, ny=value.ny + 1, nz=value.nz + 1)
-
     cell_size: Size3d = SchemaProperty(
         "cell_size",
         TypeAdapter(Size3d),
@@ -193,52 +187,5 @@ class Regular3DGrid(BaseSpatialObject):
         TypeAdapter(Rotation | None),
     )
 
-    @classmethod
-    async def create(
-        cls,
-        context: IContext,
-        data: Regular3DGridData,
-        parent: str | None = None,
-    ) -> Self:
-        """Create a new Regular3DGrid object.
-
-        :param context: The context to use to call Evo APIs.
-        :param data: The data for the Regular3DGrid object.
-        :param parent: The parent path for the object.
-
-        :return: The created Regular3DGrid object.
-        """
-        return await cls._create(
-            context=context,
-            parent=parent,
-            data=data,
-        )
-
-    @classmethod
-    async def replace(
-        cls,
-        context: IContext,
-        reference: str,
-        data: Regular3DGridData,
-    ) -> Self:
-        """Replace an existing Regular3DGrid object.
-
-        :param context: The context to use to call Evo APIs.
-        :param reference: The reference of the object to replace.
-        :param data: The data for the Regular3DGrid object.
-
-        :return: The new version of the Regular3DGrid object.
-        """
-        return await cls._replace(
-            context=context,
-            reference=reference,
-            data=data,
-        )
-
     def compute_bounding_box(self) -> BoundingBox:
         return _calculate_bounding_box(self.origin, self.size, self.cell_size, self.rotation)
-
-    def _reset_from_object(self) -> None:
-        super()._reset_from_object()
-        self.cells.size = self.size
-        self.vertices.size = Size3i(nx=self.size.nx + 1, ny=self.size.ny + 1, nz=self.size.nz + 1)
